% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadb.R
\name{create_metadb}
\alias{create_metadb}
\title{create _metadb}
\usage{
create_metadb(db, valdoc)
}
\arguments{
\item{db}{longolite db connection}

\item{valdoc}{json validation doc}
}
\value{
returns erro message if function fails
}
\description{
Creates meta data collection using a validation doc.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 create_metadb(db,valdoc)
 }
}
}
