% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_database.R
\name{drop_database}
\alias{drop_database}
\title{drop_database}
\usage{
drop_database(database, creds)
}
\arguments{
\item{database}{database name}

\item{creds}{RDS credentials object}
}
\value{
Returns message from mongo if sucessful, error message if not sucsessful
}
\description{
Completly drops a database and all its collections
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 d <- drop_database(lims_database,creds)

expect_equal(d$ok,1)
 }
}
}
